Modify dtypes to account for incorrect conversion from integer.

+++ lib/r_grow_distance.py
@@ -58,7 +58,7 @@ def r_grow_distance(input_raster, grass_workspace):
     gscript.run_command('r.out.gdal', flags = 'cf', input = proximity_grass_name, output = output_proximity_path, format = 'GTiff', quiet = True, type = 'Float32', createopt = 'COMPRESS=LZW')
     allocation_filename = input_raster_name + '_allo.tif'
     output_allocation_path = os.path.join(input_raster_directory, allocation_filename)
-    gscript.run_command('r.out.gdal', flags = 'cf', input = allocation_grass_name, output = output_allocation_path, format = 'GTiff', quiet = True, type = 'Float32', createopt = 'COMPRESS=LZW')
+    gscript.run_command('r.out.gdal', flags = 'cf', input = allocation_grass_name, output = output_allocation_path, format = 'GTiff', quiet = True, type = 'Float64', createopt = 'COMPRESS=LZW')
     
     # Close down temporary session and remove temporary workspace.
     temporary_session.close()
